refactor(core): break doctor/trust, procedural-*, and config/wikilink import cycles
Dependency inversion at the type level: DoctorSeverity, DoctorIssue,
TrustVerdict, InstructionFileCeilingWarning, ProceduralEntryKind,
ProceduralOutcome, ProceduralMemoryEntry move to the leaf module
src/core/brain/types.ts; LinkOutputFormat moves from config.ts to its
only consumer wikilink.ts. Adds the structural refactor plan doc.
