+++ src/core/brain/trust/instruction-file-ceiling.ts
@@ -17,7 +17,7 @@
 import { existsSync, readFileSync, statSync } from "node:fs";
 import { join } from "node:path";
 
-import type { InstructionFileCeilingWarning } from "../doctor.ts";
+import type { InstructionFileCeilingWarning } from "../types.ts";
 
 const TRACKED_FILES: ReadonlyArray<string> = Object.freeze(["CLAUDE.md", "AGENTS.md", "GEMINI.md"]);
