+++ src/core/brain/trust/operator-summary.ts
@@ -20,21 +20,24 @@
 
 import { existsSync, readdirSync, statSync } from "node:fs";
 
-import type { DoctorIssue } from "../doctor.ts";
-import { runDoctor, type RunDoctorResult, type TrustVerdict } from "../doctor.ts";
+import { runDoctor, type RunDoctorResult } from "../doctor.ts";
 import type { DreamRunSummary } from "../dream.ts";
 import { collectMaintenanceActions } from "../maintenance/collect.ts";
 import type { ActionItem } from "../maintenance/action-scorer.ts";
 import { brainDirs } from "../paths.ts";
 import { BRAIN_GUARDRAIL_DEFAULTS } from "../policy.ts";
-import type { ResolvedBrainGuardrailConfig } from "../types.ts";
+import type {
+  DoctorIssue,
+  InstructionFileCeilingWarning,
+  ResolvedBrainGuardrailConfig,
+  TrustVerdict,
+} from "../types.ts";
 import {
   computeVerificationDelta,
   type VerificationDeltaResult,
 } from "./compute-verification-delta.ts";
 import { computeTrustVerdict } from "./compute-trust-verdict.ts";
 import { checkInstructionFileCeiling } from "./instruction-file-ceiling.ts";
-import type { InstructionFileCeilingWarning } from "../doctor.ts";
 
 export interface OperatorSummary {
   readonly trust_verdict: TrustVerdict;
